pyproject.toml
[tool.poetry]
name = "django_project_med"
version = "0.1.0"
description = "Descrição do seu projeto"
authors = ["Lucas Alencar Miranda <seuemail@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "5.1.6"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"

[tool.poetry.dev-dependencies]
